Edit dialogs work

+++ frontend/src/app/features/devices/devices.component.ts
@@ -4,11 +4,14 @@ import { AddNewDeviceComponent } from './add-new-device/add-new-device.component
 import { DeviceService } from '../../core/services/http/device.service';
 import { DeviceRequest } from '../../core/models/device-request';
 import { CommonModule } from '@angular/common';
+import { EditDeviceComponent } from './edit-device/edit-device.component';
+import { FormsModule } from '@angular/forms';
+
 
 @Component({
   selector: 'app-devices',
   standalone: true,
-  imports: [CommonModule],
+  imports: [CommonModule, FormsModule],
   templateUrl: './devices.component.html',
   styleUrl: './devices.component.scss'
 })
@@ -20,6 +23,7 @@ export class DevicesComponent {
   deviceRequest: DeviceRequest = {
     userID: 0
   };
+  searchQuery: string = ''; 
 
   constructor(public dialog: MatDialog,
     private deviceService: DeviceService) { }
@@ -39,29 +43,53 @@ export class DevicesComponent {
       this.getAll(); // Refresh table after user is added
     });
   }
-  edit(device: any): void {
-    const deviceId=device.id;
-    this.deviceService.updateDevice(this.deviceRequest,deviceId).subscribe(() => {
-      console.log('Device updated successfully');
-      this.getAll();
+
+  editDialog(device: any): void {
+    const dialogRef = this.dialog.open(EditDeviceComponent, {
+      disableClose: true ,
+      data: { device: device }
     });
+    console.log(device)
+    dialogRef.afterClosed().subscribe(result => {
+      console.log('The dialog was closed');
+    });
+    dialogRef.componentInstance.deviceEdited.subscribe(() => {
+      this.getAll(); // Refresh table after user is added
+    });
+
   }
 
   delete(device:any, event: Event): void {
     console.log(device);
     const deviceId=device.deviceID;
     event.preventDefault();
-    this.deviceService.deleteDevice(deviceId).subscribe(() => {
-      this.deviceDeleted.emit();
-      console.log('Device deleted successfully');
-      this.getAll();
-    });
 
+    const confirmDelete = window.confirm('Are you sure you want to delete this device?');
+
+    if(confirmDelete) {
+      this.deviceService.deleteDevice(deviceId).subscribe(() => {
+        this.deviceDeleted.emit();
+        console.log('Device deleted successfully');
+        this.getAll();
+      });
+    }
+
+  }
+  
+  filterDevices(): void {
+    this.deviceService.getCompanyDevices().subscribe(devices => {
+      this.devices = devices.filter(device => 
+        device.deviceName.toLowerCase().startsWith(this.searchQuery.toLowerCase()) ||
+        device.reference.toLowerCase().startsWith(this.searchQuery.toLowerCase()) ||
+        device.user.company.companyName.toLowerCase().startsWith(this.searchQuery.toLowerCase())
+      );
+    });
   }
 
   getAll(): void {
     this.deviceService.getCompanyDevices().subscribe(devices => {
       this.devices = devices;
+      this.filterDevices();
     });
   }
 }
