+++ frontend/src/app/features/users/users.component.ts
@@ -6,6 +6,7 @@ import { MatDialog } from '@angular/material/dialog';
 import { AddNewUserComponent } from './add-new-user/add-new-user.component';
 import { UserRequest } from '../../core/models/user-request';
 import { UserService } from '../../core/services/http/user.service';
+import { EditUserComponent } from './edit-user/edit-user.component';
 
 @Component({
   selector: 'app-users',
@@ -21,6 +22,7 @@ export class UsersComponent {
   modalVisible: boolean = false;
   users: any[] = [];
   adminId: number = 2;
+  searchQuery: string = ''; 
 
   userRequest: UserRequest = {
   }
@@ -45,28 +47,54 @@ export class UsersComponent {
 
   }
 
-  edit(user:any): void{
-    const userId = user.id;
-    this.userService.updateUser(this.userRequest, userId).subscribe(() => {
-      console.log("User updated successfully!");
-      this.getAll();
+  editDialog(user: any): void {
+    const dialogRef = this.dialog.open(EditUserComponent, {
+      disableClose: true ,
+      data: { user: user }
     });
+    console.log(user)
+    dialogRef.afterClosed().subscribe(result => {
+      console.log('The dialog was closed');
+    });
+    dialogRef.componentInstance.userEdited.subscribe(() => {
+      this.getAll(); // Refresh table after user is added
+    });
+
   }
   delete(user:any, event: Event): void {
     console.log(user)
     const userId = user.userID;
     event.preventDefault();
-    this.userService.deleteUser(userId).subscribe(() => {
-      this.userDeleted.emit();
-      console.log('User deleted successfully');
-      this.getAll();
-    });
+
+    const confirmDelete = window.confirm('Are you sure you want to delete this user?');
+  
+    if (confirmDelete) {
+      this.userService.deleteUser(userId).subscribe(() => {
+        this.userDeleted.emit();
+        console.log('User deleted successfully');
+        this.getAll();
+      });
+    }
 
   }
 
-  getAll(): void {
+  
+  filterUsers(): void {
     this.userService.getCompanyUsers().subscribe(users => {
-      this.users = users;
+      this.users = users.filter(user => 
+        user.name.toLowerCase().startsWith(this.searchQuery.toLowerCase()) ||
+        user.surname.toLowerCase().startsWith(this.searchQuery.toLowerCase()) ||
+        user.email.toLowerCase().startsWith(this.searchQuery.toLowerCase()) ||
+        user.phoneNumber.toLowerCase().startsWith(this.searchQuery.toLowerCase())
+      );
     });
   }
-}
+  
+    getAll(): void {
+      this.userService.getCompanyUsers().subscribe(users => {
+        this.users = users;
+        this.filterUsers();
+      });
+    }
+  
+  }
